Lab_2/requirements.txt

Flask
requests
Flask-Limiter
PyJWT
python-dotenv
